+++ tests/integration_tests/network/test_websocket.py
@@ -40,9 +40,9 @@ async def test_connect_and_disconnect(websocket_server):
     client = await WebSocketClient.connect(config)
 
     # Act, Assert
-    await eventually(lambda: client.is_alive())
+    await eventually(lambda: client.is_active())
     await client.disconnect()
-    await eventually(lambda: not client.is_alive())
+    await eventually(lambda: not client.is_active())
 
 
 @pytest.mark.asyncio()
@@ -51,7 +51,7 @@ async def test_client_send_recv(websocket_server):
     store = []
     config = WebSocketConfig(_server_url(websocket_server), store.append, [])
     client = await WebSocketClient.connect(config)
-    await eventually(lambda: client.is_alive())
+    await eventually(lambda: client.is_active())
 
     # Act
     num_messages = 3
@@ -62,7 +62,7 @@ async def test_client_send_recv(websocket_server):
 
     await eventually(lambda: store == [b"connected"] + [b"Hello-response"] * 3)
     await client.disconnect()
-    await eventually(lambda: not client.is_alive())
+    await eventually(lambda: not client.is_active())
 
 
 @pytest.mark.asyncio()
@@ -71,7 +71,7 @@ async def test_client_send_recv_json(websocket_server):
     store = []
     config = WebSocketConfig(_server_url(websocket_server), store.append, [])
     client = await WebSocketClient.connect(config)
-    await eventually(lambda: client.is_alive())
+    await eventually(lambda: client.is_active())
 
     # Act
     num_messages = 3
@@ -83,7 +83,7 @@ async def test_client_send_recv_json(websocket_server):
     expected = [b"connected"] + [b'{"method":"SUBSCRIBE"}-response'] * 3
     assert store == expected
     await client.disconnect()
-    await eventually(lambda: not client.is_alive())
+    await eventually(lambda: not client.is_active())
 
 
 @pytest.mark.asyncio()
@@ -92,7 +92,7 @@ async def test_reconnect_after_close(websocket_server):
     store = []
     config = WebSocketConfig(_server_url(websocket_server), store.append, [])
     client = await WebSocketClient.connect(config)
-    await eventually(lambda: client.is_alive())
+    await eventually(lambda: client.is_active())
 
     # Act
     await client.send(b"close")
@@ -108,7 +108,7 @@ async def test_exponential_backoff(websocket_server):
     store = []
     config = WebSocketConfig(_server_url(websocket_server), store.append, [])
     client = await WebSocketClient.connect(config)
-    await eventually(lambda: client.is_alive())
+    await eventually(lambda: client.is_active())
 
     # Act
     await client.send(b"close")
@@ -128,7 +128,7 @@ async def test_websocket_headers(websocket_server):
     client = await WebSocketClient.connect(config)
 
     # Act
-    await eventually(lambda: client.is_alive())
+    await eventually(lambda: client.is_active())
     await client.send(b"Hello")
     await asyncio.sleep(0.1)